requirements.txt
praw
openai
pyyaml
python-dotenv
requests
beautifulsoup4
ddgs
streamlit
pandas
